Move monitor mirage utils.py to core.utils.mirage
